requirements.txt
diffusers==0.34.0
transformers==4.45.0
# xformers==0.0.28.post1
accelerate
opencv-python
scipy
triton
altair
pandas
controlnet-aux
Pillow
mediapipe

huggingface-hub

aiohttp
pyyaml
importlib_metadata
peft

# --- RunPod worker ---
runpod                          
omegaconf 